bug in ConsoleProcess never removes endline
when resetting the buffer after command processing, it would not count
the end line character itself. The cmdEndline variable holds the index
of the endline character, not the count of the command up to and
including it. A simple example is an empty line, cmdEndline would be
set to 0 and when passed as the usedSoFar parameter of ConsoleResetBuffer
it would do nothing and the endline character remains in the buffer.