ConsoleIoReceive integer overflow bug
ConsoleIoReceive used a 'uint8_t i' to keep track of the number of bytes
read so far. This is checked against bufferLength, which is a uint32_t.
If bufferLength is 256 or more (it is 256 in the code now) an integer
overflow occurs in the 'i' variable and it is reset back to zero every
256 bytes read.

The buffer is not overflowed since the 'i' variable is also used as the
index, but the buffer contents are overwritten causing potential
corruption of data.

This only happens if 256 or more  bytes are read in one contigous pass,
such as from a very large source buffer or extremely fast IO source.

If an infinite supply of bytes are available or a blocking getchar is
used that never returns EOF, the function would never return.